requirements.txt
transformers
torch
python-telegram-bot==20.3